updated auth and error handling in sessions and user routes  - some routes didn't have the auth  - made userAuth-check before passing data - set server errors message to 'Internal Server Error'

+++ server/src/routes/lists.ts
@@ -3,36 +3,35 @@ import express from 'express';
 import { validate as uuidValidate } from 'uuid';
 import objectBoolFilter from '../utils/helper-function/objectBoolFilter';
 
-
 const prisma = new PrismaClient()
 const listsRouter = express.Router();
 
 //get
 listsRouter.get('/:id?', async (req, res) => {
+    const user_id = res.locals?.user?.id
+    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
+        
     const { id } = req.params
     const rules = objectBoolFilter(req.query)
-    const user_id = res.locals?.user?.id
-
-    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
 
     try {
-        if (id && !uuidValidate(id)) throw new Error("Bad ID")
-        let list = null
-        if (id) {
-            list = await prisma.lists.findUnique({ where: { ...rules, id, user_id } })
-        } else {
-            list = await prisma.lists.findMany({ where: { ...rules, user_id }, orderBy: { title: 'asc' } })
-        }
+        if (id && !uuidValidate(id)) return res.status(400).json({ message: 'Bad ID' })
+        let list = id
+            ? await prisma.lists.findUnique({ where: { ...rules, id, user_id } })
+            : await prisma.lists.findMany({ where: { ...rules, user_id }, orderBy: { title: 'asc' } })
+
         res.status(200).json(list); //needs json
     } catch (e) {
         console.log("[lists]", e)
-        res.status(500).json({ message: 'error' })
+        res.status(500).json({ message: 'Internal Server Error' })
     }
 })
 
 listsRouter.post('/', async (req, res) => {
-    const data = req.body
     const user_id = res.locals?.user?.id
+    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
+
+    const data = req.body
 
     try {
         const list = await prisma.lists.create({ data: { ...data, user_id } })
@@ -40,14 +39,16 @@ listsRouter.post('/', async (req, res) => {
         console.log("[lists] Inserted:", data.title)
     } catch (e) {
         console.log("[lists]", e)
-        res.status(500).json({ message: 'error' })
+        res.status(500).json({ message: 'Internal Server Error' })
     }
 })
 
 // DELETE
 listsRouter.delete('/', async (req, res) => {
-    const { body }: { body: string[] /* lists.id[] */ } = req.body;
     const user_id = res.locals?.user?.id
+    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
+
+    const { body }: { body: string[] /* lists.id[] */ } = req.body;
 
     try {
         await prisma.lists.deleteMany({
@@ -58,18 +59,18 @@ listsRouter.delete('/', async (req, res) => {
         res.status(200).json({ message: 'Lists Deleted' });
     } catch (e) {
         console.log("[lists]", e)
-        res.status(500).json({ message: 'error' })
+        res.status(500).json({ message: 'Internal Server Error' })
     }
 })
 
 
 //patch
 listsRouter.patch('/:id', async (req, res) => {
+    const user_id = res.locals?.user?.id
+    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
+
     const { id } = req.params;
     const changes = req.body;
-    const user_id = res.locals?.user?.id
-
-    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
 
     try {
         if (!uuidValidate(id)) throw new Error("Bad ID")
@@ -82,19 +83,19 @@ listsRouter.patch('/:id', async (req, res) => {
         res.status(200).json(list);
     } catch (e) {
         console.log("[lists]", e)
-        res.status(500).json({ message: 'error' })
+        res.status(500).json({ message: 'Internal Server Error' })
     }
 })
 
 
 // ## PATCH, change the values for group of list
 listsRouter.patch('/group', async (req, res) => {
+    const user_id = res.locals?.user?.id
+    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
+
     const data = req.body; //the json only contain what changed therfore it represents 'changes'
     const { id, ...restData }: { id: string[], changes: lists } = data
     const listIDs: { id: string }[] = id.map(idvalue => ({ id: idvalue }))
-    const user_id = res.locals?.user?.id
-
-    if (!user_id) return res.status(401).json({ message: 'Unauthorized' })
 
     try {
         await prisma.lists.updateMany({
@@ -105,7 +106,7 @@ listsRouter.patch('/group', async (req, res) => {
         res.status(200).json({ message: 'Lists Edited' });
     } catch (e) {
         console.log("[lists]", e)
-        res.status(500).json({ message: 'error' })
+        res.status(500).json({ message: 'Internal Server Error' })
     }
 })
